requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
lxml==4.9.2
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==1.0.0
requests==2.28.2
rsa==4.9
soupsieve==2.4
Telethon==1.27.0
urllib3==1.26.14
